Create playbook-file.yaml
This program uses the netstat command to gather information about network interfaces and their traffic statistics. The output is then parsed using awk to extract the interface name and the number of broadcast packets. The result is filtered to remove interfaces with no broadcast traffic.

The parsed data is then passed to a query using json_query filter, which filters out interfaces with broadcast traffic below a threshold (1000 in this case). The interfaces with high broadcast traffic are stored in a variable high_broadcast_interfaces

The final task in the playbook will be to check if there is any interface with high broadcast traffic in the high_broadcast_interfaces variable and if yes, it will alert with a message.

This is just a basic example and can be modified to suit specific requirements.

Please note that this is a example and it might not work as is as it is not tested.